Add union and intersection operations
